requirements.txt
# Install with: pip install -r requirements.txt
# For a smaller set of direct deps only: pip install -e .

bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.6
iniconfig==2.3.0
invoke==2.2.1
markdown-it-py==4.0.0
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==9.0.0
packaging==26.0
paramiko==4.0.0
pluggy==1.6.0
pycparser==3.0
Pygments==2.19.2
PyNaCl==1.6.2
pyserial==3.5
pytest==9.0.2
PyYAML==6.0.3
rich==14.3.3
ruamel.yaml==0.19.1
scp==0.15.0
six==1.17.0
textfsm==2.1.0
